requirements/production.txt
# PRECAUTION: avoid production dependencies that aren't in development

-r ./base.txt

# Banco de dados
# ------------------------------------------------------------------------------
psycopg2==2.9.5 --no-binary psycopg2

# Django
# ------------------------------------------------------------------------------
django-anymail[mailgun]==7.0.0
gunicorn==20.0.4
